chore: move EcOpBuiltinRunner to own file
